Move jest tests to their own package
