+++ packages/remix-validated-form/package.json
@@ -42,12 +42,10 @@
   "devDependencies": {
     "@remix-run/react": "^1.0.6",
     "@remix-run/server-runtime": "^1.0.6",
-    "@types/jest": "^27.0.3",
     "@types/lodash": "^4.14.178",
     "@types/react": "^17.0.37",
     "fetch-blob": "^3.1.3",
     "react": "^17.0.2",
-    "ts-jest": "^27.1.1",
     "tsconfig": "*",
     "typescript": "^4.5.3"
   },
